Fix detection of LIDs that are included in other LIDs
When a LID is included in another LID the one that includes it should be
preferred unless the included LID has a Platforms: property, in which case it
is considered useable. This logic only applies when the two LIDs are in the
same directory, since otherwise it's more likely the two LIDs are simply
conflicting libraries with the same name.