Implement "minimal version selection" for dependencies
The algorithm used here is based on my understanding of
https://research.swtch.com/vgo-principles, which can be very roughly
summarized as "providing repeatable builds by preferring the lowest
possible specified version of any package".

Complex dependency constraints are dropped completely. Each package
specifies only a minimum acceptable version. If multiple packages
depend on P then the maximum minimal version of P is used. If two
packages require different major versions of P, it is an error.

A set of "active" packages may be passed to resolve-deps. No version
of an active package is ever included in the result. The idea being
that the active packages are usually checked out on a branch in the
user's workspace.