Change the way active packages are handled
Rather than listing active packages in pkg.json we look for `*/pkg.json` in the
workspace directory. Instead of downloading active packages for the user they
must use `git clone` manually instead. This actually simplifies the workflow
for the user since they already know how to clone repos but don't necessarily
know the format of `workspace.json`.

It also moves towards the package manager being agnostic about which source
control system is in use.

Also renamed find-workspace to load-workspace to better reflect that it loads a
file from disk as opposed to just finding an object in memory.