New "v1" catalog format
This separates out the catalog into one file per package and a directory
structure that is more scalable. It also reduces the possibility of conflicts
when updating any given package. A conversion tool from the old format to v1 is
included, and will soon be deleted; I just want to get it into the git history.

At the same time, some attributes are moved from `<package>` to `<release>`
since they can potentially change with each release.