Add testworks as a submodule
I was not seeing errors that fraya saw because I was using OD from the master
branch and (I am assuming) fraya was using 2020.1. Adding testworks as a
submodule should ensure that it also uses the newer version of
command-line-parser that is submoduled into this repo and I believe that solves
the problem fraya saw.