Don't write empty strings to the catalog
Some optional attributes like license-url were set to the empty string when the
package was written to the catalog. Instead don't write the attribute at all.

This also removes code that (inexplicably) threw out everything read from
pkg.json and loaded the package from the catalog if it existed in the catalog.