dylan build: prepend workspace registry to OPEN_DYLAN_USER_ROOT
If O_D_U_R isn't set, the current directory's registry is used. If O_D_U_R is
set then the current directory's registry is NOT used. `dylan build` creates
the registry in the workspace directory and depends on it being used, for
obvious reasons.

It's very useful to set
`OPEN_DYLAN_USER_ROOT=/path/to/opendylan/sources/registry` in order to get the
latest versions of libraries that are part of Open Dylan, but that didn't work
without this change and was a constant source of confusion.